refactor(backend): Split backend into deamon & cli
The deamon and the cli application now have their own subfolders. Tests
are passing and appropriate changes have been made to pyproject.toml

This is due to the requirement of the cli create and import its own
modules for subcommands.
