+++ backend/bookie_backend/bookie/main.py
@@ -1,0 +1,13 @@
+import subprocess
+
+import typer
+
+app = typer.Typer()
+
+
+@app.command(name="deamon")
+def deamon():
+    
+
+if __name__ == "__main__":
+    app()
