+++ backend/tests/test_deamon.py
@@ -4,9 +4,9 @@ from fastapi.testclient import TestClient
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from bookie_backend.data.models import Base
-from bookie_backend.deamon import app, get_db
-from bookie_backend.utils import bookmark_parser
+from bookie_backend.bookied.data.models import Base
+from bookie_backend.bookied.deamon import app, get_db
+from bookie_backend.bookied.utils import bookmark_parser
 
 path = Path("test.sqlite")
 if path.exists():
